+++ tests/test_celery.py
@@ -69,23 +69,28 @@ def test_success_queue_when_applied_synchronously(celery_app, celery_task, caplo
 def test_failure_should_log_and_call_statsd(celery_app, async_task, caplog):
     statsd_mock = celery_app.statsd_client.incr
 
-    with caplog.at_level(logging.ERROR):
+    with freeze_time() as frozen, caplog.at_level(logging.INFO):
+        async_task()
+        frozen.tick(5)
+
         async_task.on_failure(exc=Exception, task_id=1234, args=[], kwargs={}, einfo=None)
 
     statsd_mock.assert_called_once_with(f"celery.test-queue.{async_task.name}.failure")
 
-    assert f"Celery task {async_task.name} (queue: test-queue) failed" in caplog.messages
+    assert f"Celery task {async_task.name} (queue: test-queue) failed after 5.0000" in caplog.messages
 
 
 def test_failure_queue_when_applied_synchronously(celery_app, celery_task, caplog):
     statsd_mock = celery_app.statsd_client.incr
 
-    with caplog.at_level(logging.ERROR):
+    with freeze_time() as frozen, caplog.at_level(logging.ERROR):
+        celery_task()
+        frozen.tick(5)
         celery_task.on_failure(exc=Exception, task_id=1234, args=[], kwargs={}, einfo=None)
 
     statsd_mock.assert_called_once_with(f"celery.none.{celery_task.name}.failure")
 
-    assert f"Celery task {celery_task.name} (queue: none) failed" in caplog.messages
+    assert f"Celery task {celery_task.name} (queue: none) failed after 5.0000" in caplog.messages
 
 
 def test_call_exports_request_id_from_headers(mocker, request_id_task):